requirements.txt
fastapi>=0.85.0
python-decouple>=3.6
simplejson>=3.17.6
simple_rest_client==1.1.3
loguru>=0.6.0
python-dotenv>=0.21.0
fastapi_utils>=0.2.1
asgi-correlation-id>=3.0.1
starlette_exporter>=0.14.0
celery>=5.2.7
jose>=1.0.0
passlib>=1.7.4
python-jose[cryptography]>=1.16.0
uvicorn>=0.18.3
python-multipart>=0.0.5
requests>=2.28.1
aiomysql>=0.1.1
aiosqlite>=0.17.0`
asyncpg>=0.26.0
sqlalchemy>=1.4.41
sqlmodel>=0.0.8
gunicorn>=20.1.0
hypercorn>=0.13.2
sqlalchemy-database>=0.0.10
aioredis>=2.0.1
MarkupSafe>=2.1.1
greenlet>=2.0.0
fastapi-pagination>=0.10.0
fastapi_amis_admin>=0.4.1
fastapi_user_auth>=0.4.1
pybase64>=1.2.3
jaeger-client>=4.8.0
Starlette-OpenTracing>=0.1.0
opentracing_instrumentation>=3.3.1
asgiref>=3.5.2